requirements.txt
requests
jsonschema
artifact-lint-py @ git+https://github.com/whatsacomputertho/artifact-lint-py@main